+++ sdk/rust/src/lib.rs
@@ -86,4 +86,6 @@ mod response_metadata_tests;
 
 pub mod slack;
 pub use slack::*;
+pub mod slack_operations;
+pub use slack_operations::*;
 pub use webhooks::types::{SlackWebhookData, SlackWebhookEventType, SlackWebhookPayload};